+++ packages/mpv-player/src/index.ts
@@ -1,0 +1,2 @@
+export { MpvPlayerView } from "./MpvPlayerView";
+export * from "./MpvPlayer.types";